Add uv and ruff (WIP)

+++ cogs/dumb_db.py
@@ -1,10 +1,12 @@
-import os
+import asyncio
 import gzip
+import os
 import shutil
 import tempfile
+
 import discord
-import asyncio
 from discord.ext import commands
+
 from lib.bot import TMWBot
 
 PATH_TO_DB = os.getenv("PATH_TO_DB", "data/db.sqlite3")
@@ -12,8 +14,8 @@ PATH_TO_DB = os.getenv("PATH_TO_DB", "data/db.sqlite3")
 
 def create_temporary_gzip_file():
     temp_file_path = os.path.join(tempfile.gettempdir(), "db.sqlite3.gz")
-    with open(PATH_TO_DB, 'rb') as f_in:
-        with gzip.open(temp_file_path, 'wb') as f_out:
+    with open(PATH_TO_DB, "rb") as f_in:
+        with gzip.open(temp_file_path, "wb") as f_out:
             shutil.copyfileobj(f_in, f_out)
     return temp_file_path
 
@@ -22,16 +24,21 @@ class DatabasePoster(commands.Cog):
     def __init__(self, bot: TMWBot):
         self.bot = bot
 
-    @discord.app_commands.command(name="post_db", description="Gzip the database file and post it to the channel.")
+    @discord.app_commands.command(
+        name="post_db", description="Gzip the database file and post it to the channel."
+    )
     async def post_db(self, interaction: discord.Interaction):
         await interaction.response.defer()
+        temp_file_path = ""
         try:
             temp_file_path = await asyncio.to_thread(create_temporary_gzip_file)
-            await interaction.followup.send(file=discord.File(temp_file_path, filename="db.sqlite3.gz"))
+            await interaction.followup.send(
+                file=discord.File(temp_file_path, filename="db.sqlite3.gz")
+            )
         except Exception as e:
-            raise
+            raise e
         finally:
-            if os.path.exists(temp_file_path):
+            if temp_file_path and os.path.exists(temp_file_path):
                 os.remove(temp_file_path)
